requirements.txt
yt-dlp
ffmpeg-python
numpy
faster-whisper
torch
torchaudio
coremltools
python-dotenv
streamlit
pandas
langgraph
langchain-core
google-generativeai
openai
fastapi
uvicorn
websockets
streamlink
supabase
supabase
